requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
python-jose[cryptography]
passlib
kokoro>=0.9.4
soundfile
numpy
python-dotenv
aiofiles
